Add dhcp::subnet tests

+++ spec/spec_helper.rb
@@ -18,7 +18,9 @@ RSpec.configure do |c|
   c.filter_run_excluding :broken => true
 
   c.after :each do
-    FileUtils.remove_entry_secure(@puppetdir)
+    if FileTest.exists?("#{@puppetdir}")
+      FileUtils.remove_entry_secure(@puppetdir)
+    end
   end
 
   c.module_path = File.join(File.dirname(__FILE__), '../../')
